Split permission manager into manager and getters/setters
